Lower pillars instead of omitting them entirely if they are generated too close to the mesh's geometry.
I detect proximity to nearby mesh geometry in a fairly naive way,
by raycasting out from each face of the cube (each face of the cube
emits a grid of rays, spaced at 1 ray per square unit). mweeks001
suggested leveraging this to identify the lowest point of ray
collision as the adjusted height of the pillar, and it turns out
that works pretty well.

It would be more correct to formally compute the distance from each
face of the cube to every face of the input mesh (the faces we check
against could be filtered down a bit by bounds checking), but the
math for that is trickier than simple raycasting so I'm avoiding it
for now.

The algorithm as-is works fairly well in my tests so far, but it
does still occasionally mess up and generate a pillar that's too
tall for its position (probably due to rays missing by grazing the
edges of the input mesh's faces).